(FIX) Объединил хэндлеры, добавил забытую аннотацию
