--- src/main/java/ru/simbirsoft/warehouse_management/controller/advice/ValidationExceptionAdvice.java
@@ -1,33 +1,0 @@
-package ru.simbirsoft.warehouse_management.controller.advice;
-
-import org.springframework.http.HttpHeaders;
-import org.springframework.http.HttpStatus;
-import org.springframework.http.ResponseEntity;
-import org.springframework.validation.FieldError;
-import org.springframework.web.bind.MethodArgumentNotValidException;
-import org.springframework.web.bind.annotation.ControllerAdvice;
-import org.springframework.web.bind.annotation.ExceptionHandler;
-import org.springframework.web.bind.annotation.ResponseStatus;
-import org.springframework.web.context.request.WebRequest;
-import org.springframework.web.servlet.mvc.method.annotation.ResponseEntityExceptionHandler;
-
-import java.util.*;
-import java.util.stream.Collectors;
-
-@ControllerAdvice
-public class ValidationExceptionAdvice {
-
-  @ExceptionHandler(MethodArgumentNotValidException.class)
-  protected ResponseEntity<Object> handleMethodArgumentNotValid(
-      MethodArgumentNotValidException ex) {
-    Map<String, Object> body = new LinkedHashMap<>();
-    body.put("timestamp", new Date());
-    // Get all errors
-    List<String> errors =
-        ex.getBindingResult().getFieldErrors().stream()
-            .map(x -> x.getDefaultMessage())
-            .collect(Collectors.toList());
-    body.put("errors", errors);
-    return new ResponseEntity<>(body, HttpStatus.BAD_REQUEST);
-  }
-}
